requirements.txt
absl_py==1.3.0
keras==2.9.0
keras_tuner==1.1.3
tensorflow==2.9.2
tensorflow_model_analysis==0.41.1
tensorflow_transform==1.10.1
tfx==1.10.0
scikit-learn